requirements.txt
altair==5.5.0
altgraph==0.17.4
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.41.0
numpy==2.2.6
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
pip==25.0.1
plotly==6.1.2
protobuf==6.31.0
pyarrow==20.0.0
pydeck==0.9.1
pygame==2.6.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
